% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polaroid.R
\name{polaroid}
\alias{polaroid}
\title{polaroid}
\usage{
polaroid()
}
\value{
None
}
\description{
run polaroid shiny application in appDir
}
\examples{
if (interactive()) {
  polaroid()
}

}
\seealso{
https://github.com/jhk0530/polaroid
}
